Mantle: apply Xcode 12 suggested fixes.
Tests are now testing archiving and archiving with secure coding,
except for archiving and unarchiving of old MTLModel format that
doesn't support secure coding.

Mantle supports the old MTLModel format, which means archives created by
older versions of Mantle (before the `MTLModel+NSCoding` interface
existed).

The unarchiving of this format in Mantle's code is done only when we
don't use secure coding. Thus the tests that test this format are not
using secure coding.